Refactor print_content_eScreen to take string input
Move the logic of the Screen print_content_eScreen()
function to the Manager print_file_eManager() function.

Marginal bugfix. If a file is more than 100 lines
long when first opened, the contents are not displayed.

This is because in the Manager's process_ENTER_input()
function, the call to create the file window contains
the number of lines and not the number of digits in the
number of lines.